+++ elir_line_control/package.xml
@@ -49,13 +49,15 @@
   <!-- Use doc_depend for packages you need only for building documentation: -->
   <!--   <doc_depend>doxygen</doc_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>dynamixel_msgs</build_depend>
+  <build_depend>std_msgs</build_depend>
+   <build_depend>geometry_msgs</build_depend>
   <build_depend>rospy</build_depend>
-  <build_export_depend>dynamixel_msgs</build_export_depend>
+  <build_export_depend>std_msgs</build_export_depend>
+  <build_export_depend>geometry_msgs</build_export_depend>
   <build_export_depend>rospy</build_export_depend>
-  <exec_depend>dynamixel_msgs</exec_depend>
+  <exec_depend>std_msgs</exec_depend>
   <exec_depend>rospy</exec_depend>
-
+  <exec_depend>geometry_msgs</exec_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
