+++ elir_state/launch/robot_state.launch
@@ -1,7 +1,7 @@
 <launch>
-
+<!-- Loads robot description -->
   <param name="robot_description" command="$(find xacro)/xacro.py '$(find elir_description)/urdf/elir.urdf.xacro'" />
-
+<!-- Starts the state publisher -->
   <node name="State_Publisher" pkg="elir_state" type="state_publisher.py" /> 
 
   <node name="joint_state_publisher" pkg="joint_state_publisher" type="joint_state_publisher">
